OPENCAST-1332 Camera moving status can be misleading
If the camera is moving very slowly and position updates are being requested frequently,
it's possible that two subsequent position updates could return the same position, although
the camera has not stopped.

Therefore, we only regard the camera as stopped when its position has been unchanged for > 500ms.

The camera move/stop position is reflected in the UI, and also used in the command logic. We
therefore also check if the last command is a stop command, and don't repeat stop commands
prior to the camera officially having stopped moving (500ms elapsed).