OPENCAST-1373 LS-165
Initial changes for Intel Sky Lake GPU

Changes BGRA_UINT8 to RGBA_UINT8 which is more widely supported

Changes INTENSITY_UINT8 to LUMINANCE_UNINT18, but change detection is not working yet.

OPENCAST-1373 Use RGBA instead of INTENSITY or LUMINANCE

Do not specify the local workgroup size in the template calls to
avoid invalidworkgroupsize errors.

This works, templates are populated and updated, but not actually
matching targets for some reason.

LS-165 Workaround for differences between Nvidia GPUs with 1024 workgroups and
Intel GPUS with 256 workgroups.

On Intel GPUs (256 workgroups), passing the localWorkSizes parameter to enqueueNDRange() produces
an InvalidWorkGroupSize error, whereas on NVIDIA GPUs (1024 workgroups), this parameter must be passed,
otherwise you get an InvalidWorkGroupSize error.

Unclear whether this is related to the number of workgroups available, or the OpenCL versions, or the
driver versions.